chore: added code to check if user if authenticated

+++ controllers/feed.js
@@ -126,6 +126,11 @@ export const updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -154,12 +159,22 @@ export const deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      // todo: check logged in user
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authenticated!");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
